Move these configuration options into conftest.py
