+++ acis_thermal_check/regression_testing.py
@@ -6,18 +6,7 @@ import shutil
 import numpy as np
 import tempfile
 from .main import ACISThermalCheck
-import pytest
 import six
-
-def pytest_addoption(parser):
-    parser.addoption("--answer_store",
-                     help="Generate new answers, but don't test. "
-                          "Argument is the directory to store the answers to.")
-
-@pytest.fixture()
-def answer_store(request):
-    return request.config.getoption('--answer_store')
-
 
 months = ["JAN", "FEB", "MAR", "APR", "MAY", "JUN",
           "JUL", "AUG", "SEP", "OCT", "NOV", "DEC"]
